Add Toggle and Door scripts with a demo scene
The Toggle script can be used for any Lever like behaviour
The Door script really describes a more general type of behaviour in which, when a set of Toggles all become active, a function, set in the editor, will be run in response.
The demo scene illustrates a basic use case of these two classes.